+++ hypervisor/include/arch/x86/sbl_seed_parse.h
@@ -1,0 +1,12 @@
+/*
+ * Copyright (C) 2018 Intel Corporation. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#ifndef SBL_SEED_PARSE_H_
+#define SBL_SEED_PARSE_H_
+
+bool sbl_seed_parse(struct vm *vm, char *cmdline, char *out_arg, uint32_t out_len);
+
+#endif /* SBL_SEED_PARSE_H_ */
